+++ tools/jetbrains/be-conductor-plugin/src/main/java/com/somniacs/beconductor/toolwindow/SessionListPanel.java
@@ -555,7 +555,7 @@ public class SessionListPanel extends JPanel {
                 manager.createLocalShellWidget(workingDir, name);
 
         try {
-            widget.executeCommand("be-conductor attach \"" + name + "\" && exit");
+            widget.executeCommand(com.somniacs.beconductor.TerminalCommandUtil.exitOnSuccess("be-conductor attach \"" + name + "\""));
         } catch (java.io.IOException ex) {
             LOG.warn("be-conductor: failed to attach in terminal", ex);
         }